pyproject.toml
[project]
name = "osmdiff"
description = "A read-only API to OpenStreetMap diffs"
version = "0.3.2"
authors = [{ name = "Martijn van Exel", email = "m@rtijn.org" }]
dependencies = [
    "shapely>=2.0.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.31.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

keywords = ["openstreetmap", "osm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
development = [
    "pytest>=8.1.1",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
